requirements.txt
Django==4.2
djangorestframework==3.15.2
psycopg2==2.9.9